backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Pillow==11.0.0
drf-yasg==1.21.8